Add support for oauth login
Implements a basic oauth provider for MCP login. Adds a new command `opencode
auth mcp` to authenticate dynamically to MCPs. On `login` initiates an oauth
login flow and completes it with a local server endpoint. Stores keys under a
local secret store under `~/.local/share/opencode/mcp-auth` for each of the MCPs
in the users config. Then adds the needed bearer tokens to the needed remote MCP
calls.

This is very similar to the implementation in `mcp-remote`, but I wanted to
include it here for easier token renewal. At the start of the day I just need
to run `opencode auth mcp login <mcp>` for each MCP I care about.

chore: format code